Dice Roll
Added a simple dice roll program which takes the number of dice a user wants to roll and how many times. The numbers will both be validated to determine if they are numbers and valid dice and the user will be prompted with a program restart if they want, or the program will close.